Add initial layout and add some rough content
Ideally the document should have a coherent narrative but I've yet to figure what that is. For example, is it simply a reference of techniques & strategies? Is it an "opiniontated design" I consider to be good? Or something else?

These questions can maybe be answered once I've got everything out of my head.